revert changes to create page
